Add new API to build with preprocessors
- Add a new API to generate with preprocessors applied
- Add an example to build TRPL

This change utilize mdBook API to build with preprocessors.
Not only built-in `LinkPreprocessor` that handle `{{# include}}` and
`{{# rustdoc_include}}` helpers, but also 3rd party ones can be used
to preprocess before rendering the book like using mdBook. In
standalone mode, an additional option, -p or --preprocess, is added to
activate this behavior.